+++ source/Energinet.DataHub.MarketParticipant.Infrastructure/Persistence/Repositories/UserRoleRepository.cs
@@ -16,6 +16,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Energinet.DataHub.Core.App.Common.Security;
 using Energinet.DataHub.MarketParticipant.Domain.Model;
 using Energinet.DataHub.MarketParticipant.Domain.Model.Users;
 using Energinet.DataHub.MarketParticipant.Domain.Repositories;
@@ -31,6 +32,25 @@ public sealed class UserRoleRepository : IUserRoleRepository
     public UserRoleRepository(IMarketParticipantDbContext marketParticipantDbContext)
     {
         _marketParticipantDbContext = marketParticipantDbContext;
+    }
+
+    public async Task<IEnumerable<UserRole>> GetAllAsync()
+    {
+        var queryable = BuildUserRoleQuery();
+
+        var selectedUserRoleFunctions = queryable.SelectMany(e => e.EicFunctions.Select(r => new { r.EicFunction, r.UserRoleId, e.Id, e.Name, e.Description, e.Status }));
+
+        var userRoles = selectedUserRoleFunctions
+            .Select(r => new UserRole(
+                new UserRoleId(r.Id),
+                r.Name,
+                r.Description,
+                r.Status,
+                new List<Permission>(),
+                r.EicFunction));
+
+        var list = await userRoles.ToListAsync().ConfigureAwait(false);
+        return list;
     }
 
     public async Task<UserRole?> GetAsync(UserRoleId userRoleId)
@@ -74,8 +94,8 @@ public sealed class UserRoleRepository : IUserRoleRepository
             userRole.Name,
             userRole.Description,
             userRole.Status,
-            userRole.EicFunctions.Select(f => f.EicFunction),
-            userRole.Permissions.Select(p => p.Permission));
+            userRole.Permissions.Select(p => p.Permission),
+            userRole.EicFunctions.First().EicFunction);
     }
 
     private IQueryable<UserRoleEntity> BuildUserRoleQuery()