+++ source/Energinet.DataHub.MarketParticipant.Client/documents/release-notes/release-notes.md
@@ -1,5 +1,17 @@
 # Energinet.DataHub.MarketParticipant.Client Release notes
 
+## Version 2.14.6
+
+- Actor name in actor selection API.
+
+## Version 2.14.5
+
+- Get all user roles return type update.
+
+## Version 2.14.4
+
+- Add get all user roles API.
+
 ## Version 2.14.3
 
 - Add get user role API.